+++ source/shared_lib/sources/lua/lua_script.cpp
@@ -204,7 +204,7 @@ void * LuaArguments::getGenericData(int argumentIndex) const{
 	//}
 	else if(lua_isnumber(luaState, argumentIndex)) {
 		lua_Integer result = luaL_checkinteger(luaState, argumentIndex);
-		printf("\nGENERIC param %d is an int, %d!\n",argumentIndex,result);
+		printf("\nGENERIC param %d is an int, %d!\n",argumentIndex,(int)result);
 		return (void *)result;
 	}
 	else {
